refactor: scripts routes
