+++ src/routes/scripts/script_hash/cbor.ts
@@ -1,0 +1,41 @@
+import { FastifyInstance, FastifyRequest } from 'fastify';
+import * as QueryTypes from '../../../types/queries/scripts';
+import * as ResponseTypes from '../../../types/responses/scripts';
+import { getDbSync } from '../../../utils/database';
+import { handle404 } from '../../../utils/error-handler';
+import { SQLQuery } from '../../../sql';
+import { getSchemaForEndpoint } from '@blockfrost/openapi';
+
+async function route(fastify: FastifyInstance) {
+  fastify.route({
+    url: '/scripts/:script_hash/cbor',
+    method: 'GET',
+    schema: getSchemaForEndpoint('/scripts/{script_hash}/cbor'),
+    handler: async (request: FastifyRequest<QueryTypes.RequestParametersScriptHash>, reply) => {
+      const clientDbSync = await getDbSync(fastify);
+
+      try {
+        const { rows }: { rows: ResponseTypes.ScriptHashCbor[] } =
+          await clientDbSync.query<QueryTypes.ScriptHashCbor>(
+            SQLQuery.get('scripts_script_hash_cbor'),
+            [request.params.script_hash],
+          );
+
+        clientDbSync.release();
+
+        if (rows.length === 0) {
+          return handle404(reply);
+        }
+
+        return reply.send(rows[0]);
+      } catch (error) {
+        if (clientDbSync) {
+          clientDbSync.release();
+        }
+        throw error;
+      }
+    },
+  });
+}
+
+export default route;
